Refactoring: GridCreatingConditionTriangle::MovePolygonCommand implementation is separated from gridcreatingconditiontriangle.cpp.
